app/backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Configuration
pydantic-settings>=2.1.0

# Document processing
PyPDF2>=3.0.0
python-docx>=1.1.0

# Google ADK
google-adk>=1.0.0
google-genai>=0.3.0

# HTTP client
aiohttp>=3.9.0

# Production
gunicorn>=21.2.0
